chore: move CountWrite into btc encoding module
